fix(server): move CORS helper to utils to avoid DB side-effects in tests
Importing routes.private.ts in the unit test triggered module-level
side effects (setupAuth, DB-touching controller imports), causing an
unhandled ECONNREFUSED rejection against PostgreSQL during vitest
teardown. All 1553 tests passed but the job still exited with code 1.

Move isAllowedWebCorsOrigin into a side-effect-free utils/cors.ts and
update both routes.private.ts and the unit test to import from there.
